backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis (Optional - for caching and rate limiting)
# redis==5.0.1
# hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API Documentation
python-openapi==0.1.1

# Rate limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0

# Email Service
resend==0.6.0

# Error Tracking (Optional)
sentry-sdk[fastapi]==1.38.0

